scripts/requirements.txt
loguru==0.7.2
web3==6.5.0
requests==2.29.0
ccxt==3.1.17
python-dotenv==1.0.0
selenium==4.27.1